Update system service for IoTS
