+++ system-service/app/src/main/AndroidManifest.xml
@@ -15,8 +15,7 @@
  ~ specific language governing permissions and limitations
  ~ under the License.
 -->
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="org.wso2.emm.system.service"
-    android:sharedUserId="android.uid.system">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="org.wso2.iot.system.service" >
 
     <uses-permission android:name="android.permission.DISABLE_KEYGUARD" />
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" />
@@ -36,41 +35,42 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.MASTER_CLEAR"/>
 
-    <permission android:name="org.wso2.emm.system.service.permission.ACCESS"
-        android:protectionLevel="signature"></permission>
+    <permission
+        android:name="org.wso2.iot.system.service.permission.ACCESS"
+        android:protectionLevel="signature" />
 
-    <uses-permission android:name="org.wso2.emm.system.service.permission.ACCESS"/>
+    <uses-permission android:name="org.wso2.iot.system.service.permission.ACCESS"/>
 
-    <uses-permission android:name="org.wso2.emm.agent.permission.ACCESS"/>
+    <uses-permission android:name="org.wso2.iot.agent.permission.ACCESS"/>
 
     <application android:allowBackup="true" android:label="@string/app_name">
 
-        <service android:name="org.wso2.emm.system.service.EMMSystemService"
+        <service android:name="org.wso2.iot.system.service.SystemService"
             android:enabled="true"
-            android:permission="org.wso2.emm.system.service.permission.ACCESS"
+            android:permission="org.wso2.iot.system.service.permission.ACCESS"
             android:exported="true">
             <intent-filter>
-                <action android:name="org.wso2.emm.system.service.START_SERVICE" />
+                <action android:name="org.wso2.iot.system.service.START_SERVICE" />
             </intent-filter>
         </service>
 
         <activity
-                android:name=".MainActivity"
+                android:name="org.wso2.iot.system.service.MainActivity"
                 android:exported="true"
                 android:screenOrientation="portrait">
         </activity>
 
         <activity
-                android:name=".LockActivity"
+                android:name="org.wso2.iot.system.service.LockActivity"
                 android:screenOrientation="portrait"
                 android:configChanges="orientation|keyboardHidden|screenSize"
                 android:label="@string/title_activity_lock">
         </activity>
 
-        <receiver android:name=".services.NotificationActionReceiver" />
+        <receiver android:name="org.wso2.iot.system.service.services.NotificationActionReceiver" />
 
         <receiver
-            android:name=".services.NetworkConnectivityStatusReceiver"
+            android:name="org.wso2.iot.system.service.services.NetworkConnectivityStatusReceiver"
             android:label="NetworkConnectivityStatusReceiver" >
             <intent-filter>
                 <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
@@ -79,7 +79,7 @@
         </receiver>
 
         <receiver
-            android:name=".services.DeviceStartupIntentReceiver"
+            android:name="org.wso2.iot.system.service.services.DeviceStartupIntentReceiver"
             android:enabled="true"
             android:exported="false" >
             <intent-filter>
@@ -91,7 +91,7 @@
         </receiver>
 
         <receiver
-            android:name=".services.BatteryChargingStateReceiver"
+            android:name="org.wso2.iot.system.service.services.BatteryChargingStateReceiver"
             android:enabled="true"
             android:exported="false" >
             <intent-filter>
@@ -99,10 +99,10 @@
             </intent-filter>
         </receiver>
 
-        <receiver android:name=".services.AlarmReceiver" />
+        <receiver android:name="org.wso2.iot.system.service.services.AlarmReceiver" />
 
         <receiver
-                android:name="org.wso2.emm.system.service.ServiceDeviceAdminReceiver"
+                android:name="org.wso2.iot.system.service.ServiceDeviceAdminReceiver"
                 android:permission="android.permission.BIND_DEVICE_ADMIN">
             <intent-filter>
