+++ system-service/app/src/main/java/org/wso2/iot/system/service/api/OTAServerConfig.java
@@ -16,13 +16,13 @@
  * under the License.
  */
 
-package org.wso2.emm.system.service.api;
+package org.wso2.iot.system.service.api;
 
 import android.content.Context;
 import android.util.Log;
-import org.wso2.emm.system.service.R;
-import org.wso2.emm.system.service.utils.Constants;
-import org.wso2.emm.system.service.utils.Preference;
+import org.wso2.iot.system.service.R;
+import org.wso2.iot.system.service.utils.Constants;
+import org.wso2.iot.system.service.utils.Preference;
 
 import java.io.File;
 import java.net.MalformedURLException;
