+++ system-service/app/src/main/java/org/wso2/iot/system/service/services/AlarmReceiver.java
@@ -15,19 +15,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.emm.system.service.services;
+package org.wso2.iot.system.service.services;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.net.Uri;
 import android.util.Log;
-import org.wso2.emm.system.service.R;
-import org.wso2.emm.system.service.api.OTADownload;
-import org.wso2.emm.system.service.utils.AppUtils;
-import org.wso2.emm.system.service.utils.Constants;
-import org.wso2.emm.system.service.utils.Preference;
+import org.wso2.iot.system.service.R;
+import org.wso2.iot.system.service.api.OTADownload;
+import org.wso2.iot.system.service.utils.Constants;
+import org.wso2.iot.system.service.utils.Preference;
 
 /**
  * This class is a broadcast receiver which triggers on scheduled task timeouts.
