+++ system-service/app/src/main/java/org/wso2/iot/system/service/services/DeviceStartupIntentReceiver.java
@@ -15,17 +15,17 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.emm.system.service.services;
+package org.wso2.iot.system.service.services;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.res.Resources;
 import android.util.Log;
-import org.wso2.emm.system.service.R;
-import org.wso2.emm.system.service.utils.AlarmUtils;
-import org.wso2.emm.system.service.utils.Constants;
-import org.wso2.emm.system.service.utils.Preference;
+import org.wso2.iot.system.service.R;
+import org.wso2.iot.system.service.utils.AlarmUtils;
+import org.wso2.iot.system.service.utils.Constants;
+import org.wso2.iot.system.service.utils.Preference;
 import java.text.ParseException;
 
 /**
