+++ system-service/app/src/main/java/org/wso2/iot/system/service/services/NetworkConnectivityStatusReceiver.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.emm.system.service.services;
+package org.wso2.iot.system.service.services;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -24,10 +24,10 @@ import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.util.Log;
 
-import org.wso2.emm.system.service.R;
-import org.wso2.emm.system.service.api.OTADownload;
-import org.wso2.emm.system.service.utils.Constants;
-import org.wso2.emm.system.service.utils.Preference;
+import org.wso2.iot.system.service.R;
+import org.wso2.iot.system.service.api.OTADownload;
+import org.wso2.iot.system.service.utils.Constants;
+import org.wso2.iot.system.service.utils.Preference;
 
 /**
  * This class handles all the functionality required for monitoring device network connectivity.
