+++ system-service/app/src/main/java/org/wso2/iot/system/service/services/NotificationActionReceiver.java
@@ -1,4 +1,4 @@
-package org.wso2.emm.system.service.services;
+package org.wso2.iot.system.service.services;
 
 import android.app.NotificationManager;
 import android.content.BroadcastReceiver;
@@ -6,11 +6,11 @@ import android.content.Context;
 import android.content.Intent;
 import android.util.Log;
 
-import org.wso2.emm.system.service.R;
-import org.wso2.emm.system.service.api.OTAServerManager;
-import org.wso2.emm.system.service.utils.CommonUtils;
-import org.wso2.emm.system.service.utils.Constants;
-import org.wso2.emm.system.service.utils.Preference;
+import org.wso2.iot.system.service.R;
+import org.wso2.iot.system.service.api.OTAServerManager;
+import org.wso2.iot.system.service.utils.CommonUtils;
+import org.wso2.iot.system.service.utils.Constants;
+import org.wso2.iot.system.service.utils.Preference;
 
 import java.net.MalformedURLException;
