+++ system-service/app/src/main/res/layout/activity_main.xml
@@ -8,6 +8,6 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         android:paddingBottom="@dimen/activity_vertical_margin"
-        tools:context="org.wso2.emm.system.service.MainActivity">
+        tools:context="org.wso2.iot.system.service.MainActivity">
 
 </RelativeLayout>
